gmetad: pthread_t is an opaque type which could oveflow an int
printing a pthread_t is not valid (as it is an opaque type), but at 
least this will avoid overflows in 64bit platforms where pointer is used

+++ gmetad/data_thread.c
@@ -28,7 +28,7 @@ data_thread ( void *arg )
  
    if(get_debug_msg_level())
       {
-         fprintf(stderr,"Data thread %d is monitoring [%s] data source\n", (int) pthread_self(), d->name);
+         fprintf(stderr,"Data thread %lu is monitoring [%s] data source\n", (unsigned long)pthread_self(), d->name);
          for(i = 0; i < d->num_sources; i++)
             {
                addr = d->sources[i];
